Split backend compilation cache
This is the first step in simplifying compilation and reusing caching across backends
